+++ docs/Changelog.md
@@ -5,6 +5,7 @@
 - New options `autofit-image` and `autofit-audio`, like autofit but used for image and audio files. Default 80.
 - New [auto-mode](https://github.com/stax76/mpv-scripts) script to use mpv and mpv.net as image viewer and audio player.
 - Fix long commands causing key bindings not visible in the command palette.
+- Fix logo overlap using mordenx.lua, if the file mordenx.lua exists, the mpv.net logo is disabled.
 
 
 # 6.0.0.0 Beta (2022-06-05)