+++ docs/Changelog.md
@@ -4,8 +4,9 @@
 - New tutorial: [Extending mpv and mpv.net via Lua scripting](https://github.com/stax76/mpv.net/wiki/Extending-mpv-and-mpv.net-via-Lua-scripting)
 - New options `autofit-image` and `autofit-audio`, like autofit but used for image and audio files. Default 80.
 - New [auto-mode](https://github.com/stax76/mpv-scripts) script to use mpv and mpv.net as image viewer and audio player.
+- New support of the mpv option `snap-window`.
 - Fix long commands causing key bindings not visible in the command palette.
-- Fix compatibility with mpv-osc-tethys, which is a new osc script.
+- Fix script compatibility with mordenx and mpv-osc-tethys.
 - Fix borderless window not resizable with mouse.